backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
joblib>=1.4.0
python-dotenv>=1.0.1
pydantic>=2.7.0
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
bcrypt>=4.1.0
